These changes may be removed because the issue is becoming more involved and I may need to rethink how this will work.
