+++ HIRS_Utils/src/main/java/hirs/data/persist/SupportReferenceManifest.java
@@ -5,14 +5,12 @@ import hirs.persist.ReferenceManifestManager;
 import hirs.persist.ReferenceManifestSelector;
 import hirs.tpm.eventlog.TCGEventLog;
 import hirs.tpm.eventlog.TpmPcrEvent;
-import org.apache.commons.codec.binary.Hex;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import java.io.IOException;
-import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.cert.CertificateException;
 import java.util.ArrayList;
@@ -25,14 +23,6 @@ import java.util.Collection;
 @Entity
 public class SupportReferenceManifest extends ReferenceManifest {
     private static final Logger LOGGER = LogManager.getLogger(SupportReferenceManifest.class);
-    /**
-     * Holds the name of the 'hexDecHash' field.
-     */
-    public static final String HEX_DEC_HASH_FIELD = "hexDecHash";
-
-    @Column
-    @JsonIgnore
-    private String hexDecHash = "";
     @Column
     @JsonIgnore
     private int pcrHash = 0;
@@ -114,15 +104,6 @@ public class SupportReferenceManifest extends ReferenceManifest {
         this.setFileName(fileName);
         this.setRimType(SUPPORT_RIM);
         this.pcrHash = 0;
-        MessageDigest digest = null;
-        this.hexDecHash = "";
-        try {
-            digest = MessageDigest.getInstance("SHA-256");
-            this.hexDecHash = Hex.encodeHexString(
-                    digest.digest(rimBytes));
-        } catch (NoSuchAlgorithmException noSaEx) {
-            LOGGER.error(noSaEx);
-        }
     }
 
     /**
@@ -252,14 +233,5 @@ public class SupportReferenceManifest extends ReferenceManifest {
      */
     public boolean isBaseSupport() {
         return !this.isSwidSupplemental() && !this.isSwidPatch();
-    }
-
-    /**
-     * Getter for the Reference Integrity Manifest hash value.
-     *
-     * @return int representation of the hash value
-     */
-    public String getHexDecHash() {
-        return hexDecHash;
     }
 }
